% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoinformatics.R
\name{addsub_formulas}
\alias{addsub_formulas}
\title{Add/sub a chemical formula to another formula}
\usage{
addsub_formulas(frml1, frml2, mode = "add")
}
\arguments{
\item{frml1}{character. Chemical formula to which something is
added/subtracted.}

\item{frml2}{character. Chemical formula that is added/subtracted from
'frml1'}

\item{mode}{character. Either 'add' or 'sub'}
}
\value{
character. Chemical formula of product.
}
\description{
Add/sub a chemical formula to another formula
}
\details{
If an element is subtracted in a way, that the product has a
negative atom count, NA is returned.
}
