% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxotrophy_frequencies.R
\name{auxotrophy_frequencies}
\alias{auxotrophy_frequencies}
\title{Calculate auxotrophy frequency table}
\usage{
auxotrophy_frequencies(
  mic,
  model.dir,
  compounds = NULL,
  min.growth = 0.005,
  min.growth.fraction = 0.05,
  n.cores = NULL
)
}
\arguments{
\item{mic}{A Microbiome object.}

\item{model.dir}{Path to the directory, where gapseq model prediction files
are stored. Files should be stored in model-named subdirectories.}
}
\value{
A list. TODO: provide results
}
\description{
Bases on the ASV/OTU-Count table and the mapping of OTUs/ASVs to
models/genomes, this function calculated the relative abundancies of
auxotrophic bacteria in each sample by metabolites (e.g. amino acids).
}
