% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoinformatics.R
\name{calc_ppmerr}
\alias{calc_ppmerr}
\title{Calculate PPM error between two mz values}
\usage{
calc_ppmerr(theo.mz, meas.mz)
}
\arguments{
\item{theo.mz}{double. Theoretical MZ values}

\item{meas.mz}{double. Measured/observed MZ values}
}
\value{
double. Error in 'ppm'.
}
\description{
Calculate PPM error between two mz values (measured vs
theoretical)
}
