% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_mult_models.R
\name{communityFBA_FB}
\alias{communityFBA_FB}
\title{Community FBA with fixed biomass ratios}
\usage{
communityFBA_FB(
  models,
  model.prop,
  scale.boundaries = 1,
  cpl_c = 400,
  cpl_u = 0.01,
  pFBAcoeff = 1e-06,
  lp.method = "hybbaropt"
)
}
\arguments{
\item{scale.boundaries}{Factor (x) that scaled the lower bounds of Exchange Reactions. Experimental, thus it is recommended to use the default: 1 (no scaling).}

\item{cpl_c}{Coupling constraint \code{c}. See Heinken et al (2013) Gut Microbes (doi: 10.4161/gmic.22370). Default: 400}

\item{cpl_u}{Coupling constraint \code{u}. See Heinken et al (2013) Gut Microbes (doi: 10.4161/gmic.22370). Default: 0.01}

\item{pFBAcoeff}{parsimonious FBA coefficient that corresponds to the weight of absolute flux values relative to the biomass mass production.}

\item{lp.method}{Solver algorith to use. Default: hybbaropt.}
}
\description{
Community FBA with fixed biomass ratios
}
