% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{constrain_mod}
\alias{constrain_mod}
\title{Constrain model}
\usage{
constrain_mod(mod, media.id = NULL, media.file = NULL, nasp = "auto")
}
\arguments{
\item{mod}{Model file of class `modelorg`}

\item{media.id}{ID for available sybil.tools medium to which model is constraint to.}

\item{media.file}{Comma- or Tab-separated table with three columns: First - Compound ID; Second - Compound name; Third - maximum uptake flux.}

\item{nasp}{Namespace of the model. Default: "auto". See details for supported identifier namespaces. Has only an effect if media is selected via `media.id`}
}
\value{
A model object of class `modelorg`.
}
\description{
Constrains a model to a specific set of lower bounds for exchange reactions.
}
\details{
Three metabolite namespaces are supported: `seed` for ModelSEED, `bigg` for BIGG, and `vmh` for Virtual Metabolic Human. If `auto`, the functions takes an educated guess.
}
