% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_table.R
\name{create_model_table}
\alias{create_model_table}
\title{Transform ASV/OTU count table to model count table}
\usage{
create_model_table(object)
}
\arguments{
\item{object}{A Microbiome object.}
}
\description{
Transform ASV/OTU count table to model count table
}
