% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mapping.R
\name{filter_mapping}
\alias{filter_mapping}
\title{Filter amplicon seq. to model mappings}
\usage{
filter_mapping(
  object,
  allowed.identity.deviation = 0,
  method.resolve.multiple = "user"
)
}
\arguments{
\item{object}{A Microbiome object.}

\item{allowed.identity.deviation}{double. Allowed deviation in itentity
between best hit and subobtimal hits.}

\item{method.resolve.multiple}{Method to use, when an ASV/OTU has several
optimal hits with the same statistics. Available methods: "user", "random",
"first".}
}
\description{
Filter amplicon seq. to model mappings
}
