% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_samples.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter samples}
\usage{
filter_samples(object, min.seqs = 5000, max.unclassified = 0.3)
}
\arguments{
\item{min.seqs}{integer. Minimum count of total ASVs/OTUs per sample.}

\item{max.unclassified}{double. Maximum proportion of counts, that correspond
to unclassified ASVs/OTUs, i.e. not mapped to any reference genome/model.}

\item{mic}{A Microbiome object.}
}
\description{
Filter samples
}
