% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_mult_models.R
\name{get_coupling_constraints_mult}
\alias{get_coupling_constraints_mult}
\title{Couple organisms' internal reactions to their own biomass production}
\usage{
get_coupling_constraints_mult(modj, cpl_c = 400, cpl_u = 0.01)
}
\arguments{
\item{modj}{Community model as returned from join_mult_models()$modj}

\item{cpl_c}{Coupling parameter c}

\item{cpl_u}{Coupling parameter u}
}
\description{
Couple organisms' internal reactions to their own biomass production
}
