% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_essential_nutrients}
\alias{get_essential_nutrients}
\title{Get essential nutrients}
\usage{
get_essential_nutrients(mod, min.growth.fraction = 0.01)
}
\arguments{
\item{mod}{Model of class `modelorg`}

\item{min.growth.fraction}{Minimum required predicted growth without metabolite relative to the predicted growth with the focal metabolite, to consider the metabolite as non-essential. Default: 0.01 (i.e. 1\% of original growth)}
}
\value{
data.table with the columns: 1 - ID of exchange reaction for the focal nutrient metabolite, 2 - Name of the exchange reaction, 3 - predicted growth rate w/o metabolite, 4 - Predicted ration of growth w/o metabolite to growth w/ metabolite. 5 - Essentiality prediction.
}
\description{
Predicts which metabolites are essential for the model's growth.
}
\examples{
eure <- readRDS(system.file("extdata", "eure.RDS", package = "MicrobiomeGS2"))
dt <- get_essential_nutrients(eure)
dt

}
