% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_exchanges}
\alias{get_exchanges}
\title{Get all exchange flux values}
\usage{
get_exchanges(
  model,
  algorithm = "mtf",
  combine.compounds = list(`DL-Lactate` = c("EX_cpd00159_e0", "EX_cpd00221_e0"))
)
}
\arguments{
\item{model}{Model file of class `modelorg`, or a named-list of `modelorg`
objects.}

\item{algorithm}{Algorithm to use to calculate flux distribution. Either 'mtf'
or 'fba'.}

\item{combine.compounds}{list of character vectors. Optional. This option can
be used to add up the exchange fluxes of specific compounds. This for instance
makes sense for enantiomers such as D- and L-Lactate, which is also set as
default and example.}
}
\value{
A data.table, with the columns: 'model' for model ID, 'rxn' for the
ID if the exchange reaction, 'flux' for the predicted flux value, 'in.model'
specifying if the exchange reaction is part of the model (TRUE) or
if absent (FALSE), 'name' for metabolite names.
}
\description{
Performs FBA(-MTF) to predict fluxes of metabolite exchange
reactions
}
\details{
In case an exchange reaction is not part of one model but present in another
model of the input models, a flux value of 0 is indicated.
}
