% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoinformatics.R
\name{get_mass}
\alias{get_mass}
\title{Get monoisotopic mass}
\usage{
get_mass(formula, charge = rep(0, length(formula)))
}
\arguments{
\item{formula}{character (vector) with chemical formulas.}

\item{charge}{integer (vector) of same length as 'formula' specifying the
compounds charge.}
}
\value{
double. exact monoisotopic mass of molecule
}
\description{
Get monoisotopic mass from compounds' formula and charge
}
