% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_metabolite_production_capacity}
\alias{get_metabolite_production_capacity}
\title{Prediction of maximum production rates}
\usage{
get_metabolite_production_capacity(mod, met)
}
\arguments{
\item{mod}{Model of class \code{modelorg}}

\item{met}{Character vector of metabolite IDs, whose production capacity is predicted.}
}
\value{
A named numeric vector with individual maximum production rates. Names correspond to metabolite IDs.
}
\description{
Gets the maximum production capacity of metabolite \code{met}.
Implementation: A new reaction with the outflow of the metabolite of interest
is introduced and it's flux maximized
using FBA.
}
