% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_mod_namespace}
\alias{get_mod_namespace}
\title{Guess compound identifier namespace}
\usage{
get_mod_namespace(mod)
}
\arguments{
\item{mod}{Model of class `modelorg`}
}
\value{
Character. Either 'seed', 'bigg', or 'vhm'
}
\description{
Guesses the namespace of metabolite identifiers
}
\examples{
ecoli <- readRDS(system.file("extdata", "ecoli.RDS", package = "MicrobiomeGS2"))
get_mod_namespace(ecoli)

bilo <- readRDS(system.file("extdata", "bilo.RDS", package = "MicrobiomeGS2"))
get_mod_namespace(bilo)

}
