% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis_tools.R
\name{get_pathway_coverage}
\alias{get_pathway_coverage}
\title{Pathway coverage}
\usage{
get_pathway_coverage(
  models,
  reactions,
  pathways,
  pathways.of.interest = NULL,
  multi.thread = TRUE,
  ncores = NULL
)
}
\arguments{
\item{models}{Object of class `modelorg` or list of `modelorg` objects.}

\item{reactions}{data.table or list of data.table for gapseq predictions
(files: *-all-Reactions.tbl). See example.}

\item{pathways}{data.table or list of data.table for gapseq predictions
(files: *-all-Pathways.tbl). See example.}

\item{pathways.of.interest}{character vector of pathway IDs to be tested for
coverage. If 'NULL', all pathways in 'pathways' are considered.}

\item{multi.thread}{logical. Indicating if parallel processing of models is
used.}

\item{ncores}{integer. Number of CPUs that are used in case of parallel
processing. If NULL, the number of available CPUs is detected.}
}
\value{
Returns a data.table. See details
}
\description{
Analyses a pathway by stating which reactions are part of a
model and which not. Works only for gapseq reconstructions.
}
\details{
The output data.table has the columns: pathway (Pathway-ID),
rxn.metacyc (Reaction within pathway), rxn.name (Reaction name), spontaneous
(TRUE if reaction is non-enzymatic), rxns.models (reaction ids in the model),
prediction (TRUE if reaction is present in the model).<br>
Please note that he order of objects in models, reactions, and pathways are
required to be the same with respect to the models.
}
\examples{
mods <- fetch_model_collection("/mnt/nuuk/2021/HRGM/models/", subset = 100)
rxns <- fetch_model_collection("/mnt/nuuk/2021/HRGM/models/",
                               file.type = "reactions", subset = 100)
pwys <- fetch_model_collection("/mnt/nuuk/2021/HRGM/models/",
                               file.type = "pathways", subset = 100)

pwys_cov <- get_pathway_coverage(mods, rxns, pwys,
                                 pathways.of.interest = c("TRPSYN-PWY",
                                                          "TRPSYN-PWY2",
                                                          "HISTSYN-PWY"))

}
