% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_produced_metabolites}
\alias{get_produced_metabolites}
\title{Get produced metabolites}
\usage{
get_produced_metabolites(
  mod,
  easyConstraints = NULL,
  multi.thread = TRUE,
  ncores = NULL
)
}
\arguments{
\item{mod}{Model object of class `modelorg`}

\item{easyConstraints}{Additional constraints. See easyConstraints in the sybil
documentation.}

\item{multi.thread}{logical. Indicating if parallel processing of models is
used.}

\item{ncores}{integer. Number of CPUs that are used in case of parallel
processing. If NULL, the number of available CPUs is detected.}
}
\value{
data.table with predicted production rate and flux ranges. Columns: `ex` - exchange reaction ID,
`rxn.name` - exchange reaction name, `l` - lower bound of reaction flux, `u` - upper bound, `mtf.flux` - predicted production flux.
}
\description{
Predicts metabolite production using MTF-FBA and FVA.
}
\details{
Ranges (`l` & `u`) are based on a FVA-derived method. This method attempts to prevent cases where a
nutrient is taken up from the environment, transformed to another compound that is produced, but without any
contribution to the organism's growth rate. When providing additional
constraints, please note that these constraints are not (yet) applied for the
flux variability analysis.
}
