% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis_tools.R
\name{get_reactions_with_metabolite}
\alias{get_reactions_with_metabolite}
\title{Get reaction involving a specific metabolite}
\usage{
get_reactions_with_metabolite(mod, met)
}
\arguments{
\item{mod}{Model of class `modelorg`}

\item{met}{Metabolite ID(s) of interest.}
}
\value{
Vector with corresponding reaction IDs.
}
\description{
Gets all reactions IDs, whose reactions involve one of the
metabolites listed in `met`
}
