% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_reduced_costs}
\alias{get_reduced_costs}
\title{Get reduced costs}
\usage{
get_reduced_costs(mod)
}
\arguments{
\item{mod}{Model of class `modelorg`}
}
\value{
data.table with first columns being the reaction ID and the second column the reduced cost value.
}
\description{
Calculated all reactions' reduced costs. For the interpretation
of reduced costs values see for instance (https://dx.doi.org/10.1002%2Fbiot.201200291)
section 6.
}
