% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_utilized_metabolites}
\alias{get_utilized_metabolites}
\title{Get utilized metabolites}
\usage{
get_utilized_metabolites(mod)
}
\arguments{
\item{mod}{Model object of class `modelorg`}
}
\value{
data.table with predicted consumption rate and flux ranges. Columns: `ex` - exchange reaction ID,
`rxn.name` - exchange reaction name, `l` - lower bound of reaction flux, `u` - upper bound, `mtf.flux` - predicted utilization flux,
`flux.at.limit` - character specifying if the uptake is at it's constraint limit. "*" if at limit and "" if not at limit.
}
\description{
Predicts metabolite consumption using MTF-FBA and FVA.
}
