% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.R
\name{initialize,Microbiome-method}
\alias{initialize,Microbiome-method}
\title{Initialize a new Microbiome object}
\usage{
\S4method{initialize}{Microbiome}(
  .Object,
  uniq.table.file,
  model.mapping.file,
  sample.description.file = NULL,
  uniq.table.format = "mothur"
)
}
\arguments{
\item{uniq.table.file}{Path to OTU/ASV/ESV table.}

\item{model.mapping.file}{Path to blast m8-output of amplicon sequences to
reference 16S database (e.g. HRGM-16S, RefSeq, UHGG, ...).}

\item{sample.description.file}{Path to sample meta information table. Should
be a csv/tsv file with column names, while the first column needs to be names
"sample" and should correspond to the column names in 'uniq.table.file'. If
'NULL', sample names are taken from colnames of 'uniq.table.file', without
any additional meta info.}

\item{uniq.table.format}{File format of 'uniq.table.file'. See details for
supported formats.}
}
\value{
A new `Microbiome` object.
}
\description{
Create a new Microbiome object.
}
