% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.R
\docType{class}
\name{Microbiome-class}
\alias{Microbiome-class}
\alias{Microbiome}
\title{An S4 class to represent a Microbiome.}
\description{
An S4 class to represent a Microbiome.
}
\section{Slots}{

\describe{
\item{\code{uniq.table.file}}{character. Path to OTU/ASV/ESV table.}

\item{\code{model.mapping.file}}{character. Path to blast m8-output of amplicon
sequences to reference 16S database (e.g. HRGM-16S, RefSeq, UHGG, ...).}

\item{\code{sample.description.file}}{character. Path to sample meta information
table. Should be a csv/tsv file with column names, while the first column
needs to be names "sample" and should correspond to the column names in
'uniq.table.file'.}

\item{\code{model.mapping}}{data.table. Data table structure to store the assignment
of amplicon sequence identifiers to reference genome/model IDs.}

\item{\code{model.mapping.single}}{data.table. Same as 'model.mapping', but each
amplicpon sequence is mapped only to a single reference genome/model.}

\item{\code{model.mapping.ensemble}}{data.table TODO}

\item{\code{uniq.table}}{table. OTU/ASV/ESV count table.}

\item{\code{model.table}}{table. Aggregated reference model/genome count table.}

\item{\code{sample.description}}{data.table. Samples' meta information.}

\item{\code{copy.nr.corrected}}{logical. Indicated, whether count tables are
corrected by amplicon gene copy number.}
}}

