% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoinformatics.R
\name{neutralize_formula}
\alias{neutralize_formula}
\title{Get neutral sum formula for charged compound}
\usage{
neutralize_formula(formula, charge)
}
\arguments{
\item{formula}{character (vector) with chemical formulas.}

\item{charge}{integer (vector) of same length as 'formula' specifying the
compounds charge.}
}
\description{
Transforms a formula of a charged compound to a neutral formula
by removing adding protons.
}
