% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_compositional_matrix.R
\name{pathway_compositional_matrix}
\alias{pathway_compositional_matrix}
\title{Calculate pathway composition matrix (experimental)}
\usage{
pathway_compositional_matrix(
  mic,
  model.dir,
  abun.ex.thres = 0.5,
  count.freq = FALSE
)
}
\arguments{
\item{mic}{A Microbiome object.}

\item{model.dir}{Path to the directory, where gapseq model prediction files
are stored. Files should be stored in model-named subdirectories.}

\item{abun.ex.thres}{Fraction of models in an ensemble, that needs to be
exceeded in order to coin that the pathway is also present in the respective
ASV/OTU.}

\item{count.freq}{Logical. Indicating whether the relative pathway
composition matrix (TRUE) or the absolute count table (FALSE, default) is
returned.}
}
\value{
A list. TODO: provide results
}
\description{
In principle, this function translates the OTU/ASV count table
to a pathway count table based on the gapseq predictions of the reference
models/genomes. Please note, that here not only one reference model is used
but the ensemble of models that have mapped with the same statistics to the
corresponding OTU/ASV sequence.
}
