% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_reactions.R
\name{print_reaction}
\alias{print_reaction}
\title{Print a reaction with additional options}
\usage{
print_reaction(mod, react, use.ids = FALSE)
}
\arguments{
\item{mod}{Model of class `modelorg`}

\item{react}{Vector of reaction IDs.}

\item{use.ids}{Boolean. Indicating whether metabolite IDs should be printed
instead of metabolite names.}
}
\value{
Vector of type `character` with the individual reaction equations.
}
\description{
An extension of sybil's function \code{printReaction}.
}
