% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_model_collection.R
\name{fetch_model_collection}
\alias{fetch_model_collection}
\title{Fetch a collection of gapseq models}
\usage{
fetch_model_collection(
  model.dir,
  IDs = NULL,
  file.type = "model",
  subset = NULL
)
}
\arguments{
\item{model.dir}{character. Path to the directory that contains the models'
RDS files.}

\item{IDs}{character vector that specifies which models should be retrieved
from the collection. If `NULL` all models from the collection will be
retrieved.}

\item{file.type}{character. Select which gapseq output file should be read.
One of 'model' (gap-filled model), 'draft' (draft network), 'reactions',
'pathways', 'transporters', 'medium' (predicted growth medium).}

\item{subset}{integer. For testing purposes you can choose the maximum number
of model files to be read.}
}
\value{
A named (IDs) list of with elements of class `modelorg` (in case of
file.type is 'model' or 'draft') or elements of class `data.table` otherwise.
}
\description{
Fetches a collection of genome-scale metabolic models
reconstructed by gapseq.
}
