% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_growth}
\alias{get_growth}
\title{Get growth rate / Value of objective function}
\usage{
get_growth(mod)
}
\arguments{
\item{mod}{Model of type `modelorg`}
}
\value{
Numeric. Value of optimal value of objective function. (e.g. growth
rate)
}
\description{
Uses a simple FBA to predict the models optimal value for the
objective functions (most commonly the growth rate).
}
