% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_utilized_metabolites}
\alias{get_utilized_metabolites}
\title{Get utilized metabolites}
\usage{
get_utilized_metabolites(
  mod,
  easyConstraints = NULL,
  multi.thread = TRUE,
  ncores = NULL
)
}
\arguments{
\item{mod}{Model object of class `modelorg`}

\item{easyConstraints}{Additional constraints. See easyConstraints in the sybil
documentation.}

\item{multi.thread}{logical. Indicating if parallel processing of models is
used.}

\item{ncores}{integer. Number of CPUs that are used in case of parallel
processing. If NULL, the number of available CPUs is detected.}
}
\value{
data.table with predicted consumption rate and flux ranges. Columns: `ex` - exchange reaction ID,
`rxn.name` - exchange reaction name, `l` - lower bound of reaction flux, `u` - upper bound, `mtf.flux` - predicted utilization flux,
`flux.at.limit` - character specifying if the uptake is at it's constraint limit. "*" if at limit and "" if not at limit.
When providing additional constraints, please note that these constraints are
not (yet) applied for the flux variability analysis.
}
\description{
Predicts metabolite consumption using MTF-FBA and FVA.
}
